LLM_Application/LLM08/qa-service/requirements.txt
fastapi
uvicorn
transformers==4.40.0
torch
pydantic
python-dotenv
huggingface_hub
streamlit
requests